Re-configured the MG a bit for parallel case (that ****ing sinking blob model is a real nuisance)
